Add `:grid` option to plots
To the list-plot, bar-chart, and histogram functions, I added a `:grid`
option. The default is `false`, which preserves the current behavior of
no grid lines being drawn. If `true`, both horizontal and vertical grid
lines are drawn along with the ticks on the X and Y axes. If `:x`, only
vertical grid lines are drawn, for the ticks on the X axis; likewise, if
`:y`, only horizontal grid lines are drawn, for the ticks on the Y axis.